refactor(publish): convert publish to refactor
