+++ packages/@monorepo-utils/publish/bin/cmd.js
@@ -1,0 +1,7 @@
+#!/usr/bin/env node
+const run = require("../lib/cli").run;
+
+run().catch(error => {
+    console.error(error);
+    process.exit(1);
+});
