+++ packages/@monorepo-utils/publish/package.json
@@ -1,14 +1,22 @@
 {
   "name": "@monorepo-utils/publish",
   "version": "1.2.0",
-  "description": "npm publish script for monorepo",
+  "description": "Publish packages to npm if needed. ",
+  "license": "MIT",
   "files": [
     "publish.js",
     "utils"
   ],
-  "main": "publish.js",
+  "main": "lib/publish.js",
+  "types": "lib/publish.d.ts",
   "bin": {
     "monorepo-utils-publish": "publish.js"
+  },
+  "scripts": {
+    "build": "cross-env NODE_ENV=production tsc -p .",
+    "prepublish": "npm run --if-present build",
+    "test": "jest",
+    "watch": "tsc -p . --watch"
   },
   "dependencies": {
     "@monorepo-utils/package-utils": "^1.1.0",
@@ -20,7 +28,36 @@
     "log-update": "^2.3.0",
     "meow": "^5.0.0"
   },
+  "devDependencies": {
+    "@types/jest": "^23.3.1",
+    "@types/node": "^10.5.5",
+    "consolemock": "^1.1.0",
+    "cross-env": "^5.2.0",
+    "jest": "^23.4.2",
+    "ts-jest": "^23.1.0",
+    "typescript": "^3.0.1"
+  },
   "publishConfig": {
     "access": "public"
+  },
+  "jest": {
+    "moduleFileExtensions": [
+      "ts",
+      "tsx",
+      "js",
+      "jsx",
+      "json",
+      "node"
+    ],
+    "testRegex": "./test/.*.ts$",
+    "rootDir": ".",
+    "transform": {
+      "^.+\\.tsx?$": "ts-jest"
+    },
+    "globals": {
+      "ts-jest": {
+        "skipBabel": true
+      }
+    }
   }
 }
