+++ packages/@monorepo-utils/publish/test/fixtures/lerna.json
@@ -1,0 +1,6 @@
+{
+  "packages": [
+    "packages/*",
+    "packages/@textlint/*"
+  ]
+}
