+++ packages/@monorepo-utils/publish/test/fixtures/packages/@textlint/ast-node-types/CHANGELOG.md
@@ -1,0 +1,95 @@
+# Change Log
+
+All notable changes to this project will be documented in this file.
+See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+<a name="4.0.2"></a>
+## [4.0.2](https://github.com/textlint/textlint/compare/@textlint/ast-node-types@4.0.1...@textlint/ast-node-types@4.0.2) (2018-03-25)
+
+
+### Chores
+
+* **test:** use `ts-node-test-register` for TypeScript testing ([be746d8](https://github.com/textlint/textlint/commit/be746d8)), closes [#451](https://github.com/textlint/textlint/issues/451)
+
+
+
+
+<a name="4.0.1"></a>
+## [4.0.1](https://github.com/textlint/textlint/compare/@textlint/ast-node-types@4.0.0...@textlint/ast-node-types@4.0.1) (2018-01-18)
+
+
+
+
+**Note:** Version bump only for package @textlint/ast-node-types
+
+<a name="4.0.0"></a>
+# [4.0.0](https://github.com/textlint/textlint/compare/@textlint/ast-node-types@3.0.1...@textlint/ast-node-types@4.0.0) (2017-12-31)
+
+
+### Chores
+
+* **ast-node-types:** Make ASTNodeTypes enum ([8ae3e67](https://github.com/textlint/textlint/commit/8ae3e67))
+
+
+### BREAKING CHANGES
+
+* **ast-node-types:** TypeScript need to user it as enum
+
+
+
+
+<a name="3.0.1"></a>
+## [3.0.1](https://github.com/textlint/textlint/compare/@textlint/ast-node-types@3.0.0...@textlint/ast-node-types@3.0.1) (2017-12-25)
+
+
+### Bug Fixes
+
+* **monorepo:** fix TypeScript module resolution in monorepo ([d5df499](https://github.com/textlint/textlint/commit/d5df499))
+
+
+
+
+<a name="3.0.0"></a>
+# [3.0.0](https://github.com/textlint/textlint/compare/@textlint/ast-node-types@3.0.0-next.0...@textlint/ast-node-types@3.0.0) (2017-12-18)
+
+
+
+
+**Note:** Version bump only for package @textlint/ast-node-types
+
+<a name="3.0.0-next.0"></a>
+# [3.0.0-next.0](https://github.com/textlint/textlint/compare/@textlint/ast-node-types@2.0.0...@textlint/ast-node-types@3.0.0-next.0) (2017-12-15)
+
+
+### Bug Fixes
+
+* **ast-node-types:** Make TxtNode non-weak type ([#382](https://github.com/textlint/textlint/issues/382)) ([81c0145](https://github.com/textlint/textlint/commit/81c0145))
+
+
+
+
+<a name="1.1.2"></a>
+## [1.1.2](https://github.com/textlint/textlint/compare/@textlint/ast-node-types@1.1.1...@textlint/ast-node-types@1.1.2) (2017-05-05)
+
+
+### Bug Fixes
+
+* **docs:** README ([dd98f5c](https://github.com/textlint/textlint/commit/dd98f5c))
+
+
+
+
+<a name="1.1.1"></a>
+## [1.1.1](https://github.com/textlint/textlint/compare/@textlint/ast-node-types@1.1.0...@textlint/ast-node-types@1.1.1) (2017-05-05)
+
+
+
+
+<a name="1.1.0"></a>
+# 1.1.0 (2017-05-05)
+
+
+### Features
+
+* **ast-node-types:** add ast-node-types ([2ba6583](https://github.com/textlint/textlint/commit/2ba6583))
+* **packages:** add [@textlint](https://github.com/textlint)/ast-node-types package ([66fb289](https://github.com/textlint/textlint/commit/66fb289))
