+++ packages/@monorepo-utils/publish/test/fixtures/packages/@textlint/ast-node-types/package.json
@@ -1,0 +1,34 @@
+{
+  "name": "@textlint/ast-node-types",
+  "version": "1000.0.0-test",
+  "description": "textlint AST node type definition.",
+  "keywords": [
+    "textlint"
+  ],
+  "license": "MIT",
+  "author": "azu",
+  "files": [
+    "lib/",
+    "src/"
+  ],
+  "main": "./lib/index.js",
+  "types": "./lib/index.d.ts",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/textlint/textlint.git"
+  },
+  "scripts": {
+    "build": "cross-env NODE_ENV=production tsc -p .",
+    "clean": "rimraf lib/",
+    "prepublish": "npm run build",
+    "test": "mocha \"test/**/*.{js,ts}\""
+  },
+  "devDependencies": {
+    "cross-env": "^5.1.3",
+    "mocha": "^5.1.1",
+    "rimraf": "^2.6.2",
+    "ts-node": "^7.0.0",
+    "ts-node-test-register": "^4.0.0",
+    "typescript": "~2.8.4"
+  }
+}
