+++ packages/@monorepo-utils/publish/test/fixtures/packages/@textlint/kernel/CHANGELOG.md
@@ -1,0 +1,171 @@
+# Change Log
+
+All notable changes to this project will be documented in this file.
+See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+<a name="2.0.9"></a>
+## [2.0.9](https://github.com/textlint/textlint/compare/@textlint/kernel@2.0.8...@textlint/kernel@2.0.9) (2018-04-02)
+
+
+
+
+**Note:** Version bump only for package @textlint/kernel
+
+<a name="2.0.8"></a>
+## [2.0.8](https://github.com/textlint/textlint/compare/@textlint/kernel@2.0.7...@textlint/kernel@2.0.8) (2018-04-02)
+
+
+
+
+**Note:** Version bump only for package @textlint/kernel
+
+<a name="2.0.7"></a>
+## [2.0.7](https://github.com/textlint/textlint/compare/@textlint/kernel@2.0.6...@textlint/kernel@2.0.7) (2018-03-25)
+
+
+### Chores
+
+* **test:** use `ts-node-test-register` for TypeScript testing ([be746d8](https://github.com/textlint/textlint/commit/be746d8)), closes [#451](https://github.com/textlint/textlint/issues/451)
+
+
+
+
+<a name="2.0.6"></a>
+## [2.0.6](https://github.com/textlint/textlint/compare/@textlint/kernel@2.0.5...@textlint/kernel@2.0.6) (2018-01-27)
+
+
+### Code Refactoring
+
+* **ast-traverse:** update usage of [@textlint](https://github.com/textlint)/ast-traverse ([133ab5a](https://github.com/textlint/textlint/commit/133ab5a))
+* **plugin-markdown:** update usage of [@textlint](https://github.com/textlint)/textlint-plugin-markdown ([d34ee08](https://github.com/textlint/textlint/commit/d34ee08))
+
+
+
+
+<a name="2.0.5"></a>
+## [2.0.5](https://github.com/textlint/textlint/compare/@textlint/kernel@2.0.4...@textlint/kernel@2.0.5) (2018-01-18)
+
+
+
+
+**Note:** Version bump only for package @textlint/kernel
+
+<a name="2.0.4"></a>
+## [2.0.4](https://github.com/textlint/textlint/compare/@textlint/kernel@2.0.3...@textlint/kernel@2.0.4) (2018-01-12)
+
+
+### Bug Fixes
+
+* **kernel:** fix return type of `SourceCode#getSource` ([1b55894](https://github.com/textlint/textlint/commit/1b55894))
+* **kernel:** pass the file path to preProcess() that is on FixerProcessor ([#458](https://github.com/textlint/textlint/issues/458)) ([5b947aa](https://github.com/textlint/textlint/commit/5b947aa))
+
+
+
+
+<a name="2.0.2"></a>
+## [2.0.2](https://github.com/textlint/textlint/compare/@textlint/kernel@2.0.1...@textlint/kernel@2.0.2) (2017-12-25)
+
+
+### Bug Fixes
+
+* **kernel:** fix import path ([02d98fe](https://github.com/textlint/textlint/commit/02d98fe))
+* **monorepo:** fix TypeScript module resolution in monorepo ([d5df499](https://github.com/textlint/textlint/commit/d5df499))
+
+
+
+
+<a name="2.0.1"></a>
+## [2.0.1](https://github.com/textlint/textlint/compare/@textlint/kernel@2.0.0...@textlint/kernel@2.0.1) (2017-12-19)
+
+
+
+
+**Note:** Version bump only for package @textlint/kernel
+
+<a name="2.0.0"></a>
+# [2.0.0](https://github.com/textlint/textlint/compare/@textlint/kernel@2.0.0-next.2...@textlint/kernel@2.0.0) (2017-12-18)
+
+
+
+
+**Note:** Version bump only for package @textlint/kernel
+
+<a name="2.0.0-next.2"></a>
+# [2.0.0-next.2](https://github.com/textlint/textlint/compare/@textlint/kernel@2.0.0-next.1...@textlint/kernel@2.0.0-next.2) (2017-12-18)
+
+
+
+
+**Note:** Version bump only for package @textlint/kernel
+
+<a name="2.0.0-next.1"></a>
+# [2.0.0-next.1](https://github.com/textlint/textlint/compare/@textlint/kernel@2.0.0-next.0...@textlint/kernel@2.0.0-next.1) (2017-12-17)
+
+
+
+
+**Note:** Version bump only for package @textlint/kernel
+
+<a name="2.0.0-next.0"></a>
+# [2.0.0-next.0](https://github.com/textlint/textlint/compare/@textlint/kernel@1.0.3...@textlint/kernel@2.0.0-next.0) (2017-12-15)
+
+
+### Bug Fixes
+
+* **textlint:** fix kernel test ([#374](https://github.com/textlint/textlint/issues/374)) ([d6953cc](https://github.com/textlint/textlint/commit/d6953cc))
+* **textlint-fixer-formatter:** use paths ([734806f](https://github.com/textlint/textlint/commit/734806f))
+
+
+
+
+<a name="1.0.3"></a>
+## [1.0.3](https://github.com/textlint/textlint/compare/@textlint/kernel@1.0.2...@textlint/kernel@1.0.3) (2017-11-05)
+
+
+### Bug Fixes
+
+* **textlint:** support scoped preset module  ([#329](https://github.com/textlint/textlint/issues/329)) ([a2c8f6b](https://github.com/textlint/textlint/commit/a2c8f6b))
+
+
+
+
+<a name="1.0.2"></a>
+## [1.0.2](https://github.com/textlint/textlint/compare/@textlint/kernel@1.0.1...@textlint/kernel@1.0.2) (2017-11-03)
+
+
+
+
+**Note:** Version bump only for package @textlint/kernel
+
+<a name="1.0.0"></a>
+# [1.0.0](https://github.com/textlint/textlint/compare/@textlint/kernel@1.0.0-beta.0...@textlint/kernel@1.0.0) (2017-10-28)
+
+
+
+
+**Note:** Version bump only for package @textlint/kernel
+
+<a name="0.2.1"></a>
+## [0.2.1](https://github.com/textlint/textlint/compare/@textlint/kernel@0.2.0...@textlint/kernel@0.2.1) (2017-05-21)
+
+
+
+
+<a name="0.2.0"></a>
+# [0.2.0](https://github.com/textlint/textlint/compare/@textlint/kernel@0.1.0...@textlint/kernel@0.2.0) (2017-05-21)
+
+
+### Features
+
+* **textlint-kernel:** add `configBaseDir` option (#295) ([85dad8a](https://github.com/textlint/textlint/commit/85dad8a))
+
+
+
+
+<a name="0.1.0"></a>
+# 0.1.0 (2017-05-18)
+
+
+### Features
+
+* **textlint-kernel:** Add [@textlint](https://github.com/textlint)/kernel (#292) ([30473c3](https://github.com/textlint/textlint/commit/30473c3))
