+++ packages/@monorepo-utils/publish/test/fixtures/packages/@textlint/kernel/package.json
@@ -1,0 +1,63 @@
+{
+  "name": "@textlint/kernel",
+  "version": "1000.0.0-test",
+  "description": "textlint kernel is core logic by pure JavaScript.",
+  "keywords": [
+    "textlint"
+  ],
+  "homepage": "https://github.com/textlint/textlint/tree/master/packages/@textlint/textlint-kernel/",
+  "bugs": {
+    "url": "https://github.com/textlint/textlint/issues"
+  },
+  "license": "MIT",
+  "author": "azu",
+  "files": [
+    "bin/",
+    "lib/",
+    "src/"
+  ],
+  "main": "lib/kernel/src/index.js",
+  "typings": "lib/kernel/src/index.d.ts",
+  "directories": {
+    "test": "test"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/textlint/textlint.git"
+  },
+  "scripts": {
+    "build": "npm-run-all build:src",
+    "build:src": "cross-env NODE_ENV=production tsc -p .",
+    "clean": "rimraf out/ lib/ es_modules/",
+    "prepublish": "npm run --if-present build",
+    "test": "mocha \"test/**/*.{js,ts}\""
+  },
+  "dependencies": {
+    "@textlint/ast-node-types": "^4.0.2",
+    "@textlint/ast-traverse": "^2.0.8",
+    "@textlint/feature-flag": "^3.0.4",
+    "@types/bluebird": "^3.5.18",
+    "bluebird": "^3.5.1",
+    "debug": "^2.6.6",
+    "deep-equal": "^1.0.1",
+    "map-like": "^2.0.0",
+    "object-assign": "^4.1.1",
+    "structured-source": "^3.0.2"
+  },
+  "devDependencies": {
+    "@textlint/markdown-to-ast": "^6.0.8",
+    "@types/deep-equal": "^1.0.1",
+    "@types/mocha": "^5.2.0",
+    "@types/node": "^8.0.28",
+    "cpx": "^1.5.0",
+    "cross-env": "^5.0.5",
+    "mocha": "^5.1.1",
+    "npm-run-all": "^4.1.1",
+    "rimraf": "^2.6.2",
+    "shelljs": "^0.8.2",
+    "ts-node": "^7.0.0",
+    "ts-node-test-register": "^4.0.0",
+    "typescript": "~2.8.4",
+    "unist-util-select": "^1.5.0"
+  }
+}
