+++ packages/@monorepo-utils/publish/test/fixtures/packages/@textlint/linter-formatter/CHANGELOG.md
@@ -1,0 +1,135 @@
+# Change Log
+
+All notable changes to this project will be documented in this file.
+See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+<a name="3.0.7"></a>
+## [3.0.7](https://github.com/textlint/textlint/compare/@textlint/linter-formatter@3.0.6...@textlint/linter-formatter@3.0.7) (2018-04-02)
+
+
+
+
+**Note:** Version bump only for package @textlint/linter-formatter
+
+<a name="3.0.6"></a>
+## [3.0.6](https://github.com/textlint/textlint/compare/@textlint/linter-formatter@3.0.5...@textlint/linter-formatter@3.0.6) (2018-04-02)
+
+
+
+
+**Note:** Version bump only for package @textlint/linter-formatter
+
+<a name="3.0.5"></a>
+## [3.0.5](https://github.com/textlint/textlint/compare/@textlint/linter-formatter@3.0.4...@textlint/linter-formatter@3.0.5) (2018-03-25)
+
+
+### Chores
+
+* **test:** use `ts-node-test-register` for TypeScript testing ([be746d8](https://github.com/textlint/textlint/commit/be746d8)), closes [#451](https://github.com/textlint/textlint/issues/451)
+
+
+
+
+<a name="3.0.4"></a>
+## [3.0.4](https://github.com/textlint/textlint/compare/@textlint/linter-formatter@3.0.3...@textlint/linter-formatter@3.0.4) (2018-01-27)
+
+
+
+
+**Note:** Version bump only for package @textlint/linter-formatter
+
+<a name="3.0.3"></a>
+## 3.0.3 (2018-01-18)
+
+
+### Bug Fixes
+
+* **linter-formatter:** add `publishConfig` ([9f54a5a](https://github.com/textlint/textlint/commit/9f54a5a))
+
+
+
+
+<a name="3.0.2"></a>
+## [3.0.2](https://github.com/textlint/textlint/compare/textlint-formatter@3.0.1...textlint-formatter@3.0.2) (2018-01-12)
+
+
+
+
+**Note:** Version bump only for package textlint-formatter
+
+<a name="3.0.0"></a>
+# [3.0.0](https://github.com/textlint/textlint/compare/textlint-formatter@2.0.0...textlint-formatter@3.0.0) (2017-12-25)
+
+
+### Bug Fixes
+
+* **textlint-formatter:** define `TextLintFormatterOption` interface ([31146c7](https://github.com/textlint/textlint/commit/31146c7))
+* **textlint-formatter:** run all tests ([6d507c2](https://github.com/textlint/textlint/commit/6d507c2)), closes [#402](https://github.com/textlint/textlint/issues/402)
+
+
+### Documentation
+
+* **textlint-formatter:** Update API ([8f4901f](https://github.com/textlint/textlint/commit/8f4901f))
+
+
+### Features
+
+* **textlint-formatter:** add getFormatterList function ([25e1113](https://github.com/textlint/textlint/commit/25e1113))
+* **textlint-formatter:** add new option to cli ([30351a8](https://github.com/textlint/textlint/commit/30351a8))
+
+
+### BREAKING CHANGES
+
+* **textlint-formatter:** export `{ createFormatter }` instead of `{ default }`
+
+
+
+
+<a name="2.0.0"></a>
+# [2.0.0](https://github.com/textlint/textlint/compare/textlint-formatter@2.0.0-next.1...textlint-formatter@2.0.0) (2017-12-18)
+
+
+
+
+**Note:** Version bump only for package textlint-formatter
+
+<a name="2.0.0-next.1"></a>
+# [2.0.0-next.1](https://github.com/textlint/textlint/compare/textlint-formatter@2.0.0-next.0...textlint-formatter@2.0.0-next.1) (2017-12-18)
+
+
+
+
+**Note:** Version bump only for package textlint-formatter
+
+<a name="2.0.0-next.0"></a>
+# 2.0.0-next.0 (2017-12-15)
+
+
+### Bug Fixes
+
+* **npm:** quick fix npm ([5b20137](https://github.com/textlint/textlint/commit/5b20137))
+* **npm:** resolve issue ([d214c1c](https://github.com/textlint/textlint/commit/d214c1c))
+* **options:** fix to use `color` ([9809df4](https://github.com/textlint/textlint/commit/9809df4))
+* **pretty-error:** fix `file:line:column` ([3bcb06c](https://github.com/textlint/textlint/commit/3bcb06c))
+* **pretty-error:** fix export function ([e0d568b](https://github.com/textlint/textlint/commit/e0d568b))
+* **pretty-error:** fix FATAL ERROR: CALL_AND_RETRY_LAST Allocation failed ([dd6418d](https://github.com/textlint/textlint/commit/dd6418d))
+* **pretty-error:** fix position of column indicator ([d99b222](https://github.com/textlint/textlint/commit/d99b222))
+* **pretty-error:** remove empty line ([34fc250](https://github.com/textlint/textlint/commit/34fc250))
+* **stylish:** align table cells taking unicode chars into account ([e757c5b](https://github.com/textlint/textlint/commit/e757c5b))
+* **stylish:** narrow the width of error message when it has line breaks ([7c1aa0a](https://github.com/textlint/textlint/commit/7c1aa0a))
+* **test:** column number start with 1 ([a377817](https://github.com/textlint/textlint/commit/a377817))
+* **test:** fix pretty-errro test ([7e50453](https://github.com/textlint/textlint/commit/7e50453))
+* **textlint-formatter:** unignore lib/ ([#360](https://github.com/textlint/textlint/issues/360)) ([59488d8](https://github.com/textlint/textlint/commit/59488d8))
+
+
+### Features
+
+* **cli:** add cli for textlint-formatter ([8de0c97](https://github.com/textlint/textlint/commit/8de0c97))
+* **formatter:** add json formatter ([5642c83](https://github.com/textlint/textlint/commit/5642c83))
+* **formatter:** update formatter and add tests ([6f7778d](https://github.com/textlint/textlint/commit/6f7778d))
+* **lib:** allow to specify package name ([c9828bc](https://github.com/textlint/textlint/commit/c9828bc))
+* **preety-error:** add total output ([8a02c88](https://github.com/textlint/textlint/commit/8a02c88))
+* **pretty-error:** export prettyError ([c2726ac](https://github.com/textlint/textlint/commit/c2726ac))
+* **pretty-error:** support East Asian Width problem ([8b28994](https://github.com/textlint/textlint/commit/8b28994)), closes [#2](https://github.com/textlint/textlint/issues/2)
+* **pretty-error:** support no-color option ([614c1ea](https://github.com/textlint/textlint/commit/614c1ea))
+* **stylish:** add help to use fixer ([d0eadfc](https://github.com/textlint/textlint/commit/d0eadfc))
