+++ packages/@monorepo-utils/publish/test/fixtures/packages/@textlint/linter-formatter/package.json
@@ -1,0 +1,67 @@
+{
+  "name": "@textlint/linter-formatter",
+  "version": "1000.0.0-test",
+  "description": "textlint output formatter",
+  "homepage": "https://github.com/textlint/textlint/tree/master/packages/@textlint/linter-formatter",
+  "bugs": {
+    "url": "https://github.com/textlint/textlint/issues"
+  },
+  "license": "MIT",
+  "author": "azu",
+  "files": [
+    "lib/",
+    "bin/"
+  ],
+  "main": "lib/linter-formatter/src/index.js",
+  "types": "lib/linter-formatter/src/index.d.ts",
+  "bin": {
+    "textlint-formatter": "bin/cmd.js"
+  },
+  "directories": {
+    "test": "test/"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/textlint/textlint.git"
+  },
+  "scripts": {
+    "build": "cross-env NODE_ENV=production tsc -p .",
+    "clean": "rimraf lib/",
+    "prepublish": "npm run build",
+    "test": "mocha \"test/**/*.{js,ts}\""
+  },
+  "dependencies": {
+    "@azu/format-text": "^1.0.1",
+    "@azu/style-format": "^1.0.0",
+    "@textlint/kernel": "^2.0.9",
+    "chalk": "^1.0.0",
+    "concat-stream": "^1.5.1",
+    "js-yaml": "^3.2.4",
+    "optionator": "^0.8.1",
+    "pluralize": "^2.0.0",
+    "string-width": "^1.0.1",
+    "string.prototype.padstart": "^3.0.0",
+    "strip-ansi": "^3.0.1",
+    "table": "^3.7.8",
+    "text-table": "^0.2.0",
+    "try-resolve": "^1.0.1",
+    "xml-escape": "^1.0.0"
+  },
+  "devDependencies": {
+    "@types/mocha": "^5.2.0",
+    "@types/node": "^8.0.57",
+    "chai": "^3.5.0",
+    "cross-env": "^5.1.1",
+    "mocha": "^5.1.1",
+    "power-assert": "^1.1.0",
+    "proxyquire": "^1.7.4",
+    "rimraf": "^2.6.2",
+    "sinon": "^1.17.3",
+    "ts-node": "^7.0.0",
+    "ts-node-test-register": "^4.0.0",
+    "typescript": "~2.8.4"
+  },
+  "publishConfig": {
+    "access": "public"
+  }
+}
