+++ packages/@monorepo-utils/publish/test/fixtures/packages/@textlint/markdown-to-ast/CHANGELOG.md
@@ -1,0 +1,170 @@
+# Change Log
+
+All notable changes to this project will be documented in this file.
+See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+<a name="6.0.8"></a>
+## [6.0.8](https://github.com/textlint/textlint/compare/@textlint/markdown-to-ast@6.0.7...@textlint/markdown-to-ast@6.0.8) (2018-04-02)
+
+
+### Bug Fixes
+
+* **markdown-to-ast:** enable yaml frontmatter parse by default ([121c62f](https://github.com/textlint/textlint/commit/121c62f))
+
+
+
+
+<a name="6.0.7"></a>
+## [6.0.7](https://github.com/textlint/textlint/compare/@textlint/markdown-to-ast@6.0.6...@textlint/markdown-to-ast@6.0.7) (2018-04-02)
+
+
+### Bug Fixes
+
+* **markdown-to-ast:** upgrade remark@^9.0.0 ([8a70c0f](https://github.com/textlint/textlint/commit/8a70c0f))
+
+
+### Chores
+
+* **markdown-to-ast:** update fixtures ([ac61a54](https://github.com/textlint/textlint/commit/ac61a54))
+
+
+### Tests
+
+* **markdown-to-ast:** add failure test case ([87f218d](https://github.com/textlint/textlint/commit/87f218d))
+
+
+
+
+<a name="6.0.6"></a>
+## [6.0.6](https://github.com/textlint/textlint/compare/@textlint/markdown-to-ast@6.0.5...@textlint/markdown-to-ast@6.0.6) (2018-03-25)
+
+
+### Chores
+
+* format ([d8f44db](https://github.com/textlint/textlint/commit/d8f44db))
+
+
+
+
+<a name="6.0.5"></a>
+## [6.0.5](https://github.com/textlint/textlint/compare/@textlint/markdown-to-ast@6.0.4...@textlint/markdown-to-ast@6.0.5) (2018-01-27)
+
+
+### Code Refactoring
+
+* **ast-traverse:** update usage of [@textlint](https://github.com/textlint)/ast-traverse ([133ab5a](https://github.com/textlint/textlint/commit/133ab5a))
+
+
+
+
+<a name="6.0.4"></a>
+## 6.0.4 (2018-01-18)
+
+
+### Bug Fixes
+
+* **markdown-to-ast:** add "publishConfig" ([929b575](https://github.com/textlint/textlint/commit/929b575))
+
+
+
+
+<a name="6.0.3"></a>
+## [6.0.3](https://github.com/textlint/textlint/compare/markdown-to-ast@6.0.2...markdown-to-ast@6.0.3) (2018-01-12)
+
+
+
+
+**Note:** Version bump only for package markdown-to-ast
+
+<a name="6.0.2"></a>
+## [6.0.2](https://github.com/textlint/textlint/compare/markdown-to-ast@6.0.1...markdown-to-ast@6.0.2) (2017-12-31)
+
+
+
+
+**Note:** Version bump only for package markdown-to-ast
+
+<a name="6.0.1"></a>
+## [6.0.1](https://github.com/textlint/textlint/compare/markdown-to-ast@6.0.0...markdown-to-ast@6.0.1) (2017-12-25)
+
+
+
+
+**Note:** Version bump only for package markdown-to-ast
+
+<a name="6.0.0"></a>
+# [6.0.0](https://github.com/textlint/textlint/compare/markdown-to-ast@6.0.0-next.0...markdown-to-ast@6.0.0) (2017-12-18)
+
+
+
+
+**Note:** Version bump only for package markdown-to-ast
+
+<a name="6.0.0-next.0"></a>
+# [6.0.0-next.0](https://github.com/textlint/textlint/compare/markdown-to-ast@5.0.0...markdown-to-ast@6.0.0-next.0) (2017-12-15)
+
+
+
+
+**Note:** Version bump only for package markdown-to-ast
+
+<a name="4.0.0"></a>
+# 4.0.0 (2017-05-06)
+
+
+### Bug Fixes
+
+* **example:** fix path ([3047d10](https://github.com/textlint/textlint/commit/3047d10))
+* **example:** update ([a8e4941](https://github.com/textlint/textlint/commit/a8e4941))
+* **npm:** move debug module to dependency ([eef12f0](https://github.com/textlint/textlint/commit/eef12f0))
+* **npm:** move dev to dependecies ([8e7cd32](https://github.com/textlint/textlint/commit/8e7cd32))
+* **parse:** use thematicBreak as a map key ([b9fe7d1](https://github.com/textlint/textlint/commit/b9fe7d1))
+* **readme:** fix outdated link ([5697807](https://github.com/textlint/textlint/commit/5697807)), closes [#9](https://github.com/textlint/textlint/issues/9)
+* **remark:** fix usage of remark ([8702d24](https://github.com/textlint/textlint/commit/8702d24))
+* **test:** fix test on Windows ([956c6d1](https://github.com/textlint/textlint/commit/956c6d1))
+* **test:** rename test case ([821c650](https://github.com/textlint/textlint/commit/821c650))
+* **test:** update to mdast latest ([9eee76d](https://github.com/textlint/textlint/commit/9eee76d))
+* **travis:** fix path to shell-script ([13c0174](https://github.com/textlint/textlint/commit/13c0174))
+
+
+### Features
+
+* **ast:** enable SetextHeader test ([f980d41](https://github.com/textlint/textlint/commit/f980d41))
+* **ast:** implement Break node ([fb41948](https://github.com/textlint/textlint/commit/fb41948))
+* **ast:** implement HorizontalRule node ([a01084c](https://github.com/textlint/textlint/commit/a01084c))
+* **ast:** implement Html node ([3b94068](https://github.com/textlint/textlint/commit/3b94068))
+* **ast:** implement some node ([f6bbd54](https://github.com/textlint/textlint/commit/f6bbd54))
+* **ast:** re-implement Code ([832cbd6](https://github.com/textlint/textlint/commit/832cbd6))
+* **ast:** re-implement CodeBlock ([902a6c8](https://github.com/textlint/textlint/commit/902a6c8))
+* **ast:** re-implement Emphasis node ([6faf2f3](https://github.com/textlint/textlint/commit/6faf2f3))
+* **ast:** re-implement Strong node ([307c037](https://github.com/textlint/textlint/commit/307c037))
+* **demo:** add title ([174ac1b](https://github.com/textlint/textlint/commit/174ac1b))
+* **markdown:** update remark to 4.x ([f44a326](https://github.com/textlint/textlint/commit/f44a326))
+* **markdown-to-ast:** update remark[@7](https://github.com/7) (#278) ([197d5d4](https://github.com/textlint/textlint/commit/197d5d4))
+* **parser:** add `loc` object to Document type node ([dcc7e2b](https://github.com/textlint/textlint/commit/dcc7e2b))
+* **parser:** add workaround for Header ([e79ab46](https://github.com/textlint/textlint/commit/e79ab46))
+* **parser:** correct `range` of `CodeBlock` ([583b1d7](https://github.com/textlint/textlint/commit/583b1d7))
+* **parser:** correct `raw` property of `BlockQuote` ([a38bdf9](https://github.com/textlint/textlint/commit/a38bdf9))
+* **parser:** correct `raw` property of `Code` ([ef4535f](https://github.com/textlint/textlint/commit/ef4535f))
+* **parser:** correct `raw` property of `Emph` ([6da8e88](https://github.com/textlint/textlint/commit/6da8e88))
+* **parser:** correct `raw` property of `Image` ([27f46d0](https://github.com/textlint/textlint/commit/27f46d0))
+* **parser:** correct `raw` property of `List` type ([59f1723](https://github.com/textlint/textlint/commit/59f1723))
+* **parser:** correct `raw` property of `Strong` ([2265122](https://github.com/textlint/textlint/commit/2265122))
+* **parser:** correct `raw` property of LiteItem ([8414d50](https://github.com/textlint/textlint/commit/8414d50))
+* **parser:** correct `start_column` of each nodes. ([408e97b](https://github.com/textlint/textlint/commit/408e97b))
+* **parser:** implement Header of `raw` ([740d11a](https://github.com/textlint/textlint/commit/740d11a))
+* **parser:** remove un-used properties ([f72fde8](https://github.com/textlint/textlint/commit/f72fde8))
+* **parser:** update to commonmark 0.15 ([f31eaa3](https://github.com/textlint/textlint/commit/f31eaa3))
+* **travis:** add npm run build ([69eaa48](https://github.com/textlint/textlint/commit/69eaa48))
+* **travis:** add travis badge ([7121865](https://github.com/textlint/textlint/commit/7121865))
+* **travis:** auto deploy to gh-page ([a686595](https://github.com/textlint/textlint/commit/a686595))
+
+
+### BREAKING CHANGES
+
+* **markdown-to-ast:** some parse result is changed.
+
+It is correct, but we think that is a breacking change.
+This change have possibilities that the user update patch-version textlint and fail test.
+
+close https://github.com/textlint/markdown-to-ast/issues/14
