+++ packages/@monorepo-utils/publish/test/fixtures/packages/@textlint/markdown-to-ast/package.json
@@ -1,0 +1,58 @@
+{
+  "name": "@textlint/markdown-to-ast",
+  "version": "1000.0.0-test",
+  "description": "Parse Markdown to AST with location info.",
+  "homepage": "https://github.com/textlint/textlint/tree/master/packages/@textlint/markdown-to-ast/",
+  "bugs": {
+    "url": "https://github.com/textlint/textlint/issues"
+  },
+  "license": "MIT",
+  "author": "azu",
+  "files": [
+    "lib/",
+    "src/"
+  ],
+  "main": "lib/markdown-parser.js",
+  "directories": {
+    "test": "test/"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/textlint/textlint.git"
+  },
+  "scripts": {
+    "build": "cross-env NODE_ENV=production babel src --out-dir lib --source-maps",
+    "clean": "rimraf lib/",
+    "example:build": "browserify example/js/index.js -o example/app/app.js",
+    "prepublish": "npm run --if-present build",
+    "test": "mocha \"test/**/*.{js,ts}\"",
+    "watch": "babel src --out-dir lib --watch --source-maps"
+  },
+  "dependencies": {
+    "@textlint/ast-node-types": "^4.0.2",
+    "debug": "^2.1.3",
+    "remark-frontmatter": "^1.2.0",
+    "remark-parse": "^5.0.0",
+    "structured-source": "^3.0.2",
+    "traverse": "^0.6.6",
+    "unified": "^6.1.6"
+  },
+  "devDependencies": {
+    "@textlint/ast-tester": "^2.0.7",
+    "@textlint/ast-traverse": "^2.0.8",
+    "babel-cli": "^6.24.1",
+    "babel-preset-env": "^1.4.0",
+    "babel-preset-jsdoc-to-assert": "^4.0.0",
+    "babel-preset-power-assert": "^1.0.0",
+    "babel-register": "^6.24.1",
+    "browserify": "^14.3.0",
+    "cross-env": "^4.0.0",
+    "mkdirp": "^0.5.1",
+    "mocha": "^5.1.1",
+    "power-assert": "^1.4.2",
+    "rimraf": "^2.6.2"
+  },
+  "publishConfig": {
+    "access": "public"
+  }
+}
