+++ packages/@monorepo-utils/publish/test/fixtures/packages/textlint/package.json
@@ -1,0 +1,102 @@
+{
+  "name": "textlint",
+  "version": "1000.0.0-test",
+  "description": "The pluggable linting tool for text and markdown.",
+  "keywords": [
+    "AST",
+    "lint",
+    "linting",
+    "markdown",
+    "plugable",
+    "text",
+    "textlint"
+  ],
+  "homepage": "https://github.com/textlint/textlint/",
+  "bugs": {
+    "url": "https://github.com/textlint/textlint/issues"
+  },
+  "license": "MIT",
+  "author": "azu",
+  "files": [
+    "bin/",
+    "lib/textlint/src",
+    "src/",
+    "typings/"
+  ],
+  "main": "lib/textlint/src/index.js",
+  "types": "lib/textlint/src/index.d.ts",
+  "bin": {
+    "textlint": "./bin/textlint.js"
+  },
+  "engines": {
+    "node": ">=6.0.0"
+  },
+  "directories": {
+    "test": "test/"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/textlint/textlint.git"
+  },
+  "scripts": {
+    "build": "cross-env NODE_ENV=production tsc -p .",
+    "clean": "rimraf out/ lib/",
+    "prepublish": "npm run build",
+    "test": "npm-run-all test:unit",
+    "test:unit": "mocha \"test/**/*.{js,ts}\"",
+    "watch": "tsc -p . --watch"
+  },
+  "dependencies": {
+    "@textlint/ast-node-types": "^4.0.2",
+    "@textlint/ast-traverse": "^2.0.8",
+    "@textlint/feature-flag": "^3.0.4",
+    "@textlint/fixer-formatter": "^3.0.7",
+    "@textlint/kernel": "^2.0.9",
+    "@textlint/linter-formatter": "^3.0.7",
+    "@textlint/textlint-plugin-markdown": "^4.0.10",
+    "@textlint/textlint-plugin-text": "^3.0.10",
+    "@types/bluebird": "^3.5.18",
+    "bluebird": "^3.0.5",
+    "debug": "^2.1.0",
+    "deep-equal": "^1.0.1",
+    "file-entry-cache": "^2.0.0",
+    "get-stdin": "^5.0.1",
+    "glob": "^7.1.1",
+    "interop-require": "^1.0.0",
+    "is-file": "^1.0.0",
+    "log-symbols": "^1.0.2",
+    "map-like": "^2.0.0",
+    "md5": "^2.2.1",
+    "mkdirp": "^0.5.0",
+    "object-assign": "^4.0.1",
+    "optionator": "^0.8.0",
+    "path-to-glob-pattern": "^1.0.2",
+    "rc-config-loader": "^2.0.1",
+    "read-pkg": "^1.1.0",
+    "read-pkg-up": "^3.0.0",
+    "structured-source": "^3.0.2",
+    "try-resolve": "^1.0.1",
+    "unique-concat": "^0.2.2"
+  },
+  "devDependencies": {
+    "@types/mocha": "^5.2.0",
+    "@types/node": "^8.0.53",
+    "clone": "^2.0.0",
+    "cpx": "^1.5.0",
+    "cross-env": "^4.0.0",
+    "mocha": "^5.1.1",
+    "npm-run-all": "^4.0.2",
+    "power-assert": "^1.3.1",
+    "rimraf": "^2.6.2",
+    "shelljs": "^0.8.2",
+    "source-map-support": "^0.5.0",
+    "textlint-plugin-html": "^0.1.2",
+    "textlint-rule-helper": "^2.0.0",
+    "textlint-rule-no-todo": "^2.0.0",
+    "textlint-rule-preset-ja-spacing": "^2.0.0",
+    "textlint-rule-preset-jtf-style": "^2.3.0",
+    "ts-node": "^7.0.0",
+    "ts-node-test-register": "^4.0.0",
+    "typescript": "~2.8.4"
+  }
+}
