+++ packages/@monorepo-utils/publish/test/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../tsconfig.json",
+  "compilerOptions": {
+    "declaration": false
+  },
+  "include": [
+    "**/*"
+  ]
+}
